refactor: split checker module into elaborate and infer
Extract elaboration functions (elaborate_program, collect_signatures,
elaborate_bodies) into checker/elaborate.rs and type inference/checking
functions (infer, check, check_val) into checker/infer.rs. Keep utility
functions (builtin_prim_ty, value_type_universe_ctx, types_equal_val)
in mod.rs. This improves code organization and reduces mod.rs to ~70 LOC.
